+++ lending_app/utilities/__init__.py
@@ -1,0 +1,6 @@
+from .utility_functions import generate_short_alphanumeric_uuid, complete_months_between
+
+__all__ = [
+    'generate_short_alphanumeric_uuid'
+    'complete_months_between',
+]
